# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kynara — AI Agent Permission System
  version: 1.0.0
  summary: Authorization, auditing, and governance for AI agents and the humans they work for.
  description: |
    Kynara is an enterprise permission system for AI agents. Every tool call,
    data access, and state-changing operation your agents make can be checked
    against a central policy engine that combines RBAC, ABAC, and runtime
    context. All decisions are recorded in a tamper-evident, hash-chained audit
    log.

    ## Authentication
    - **JWT Bearer** — short-lived (15 min) access tokens issued after login or SSO.
    - **API Keys** — long-lived machine credentials, prefixed `sk_live_`, for
      agent runtimes and CI.
    - **SSO (OIDC / SAML 2.0)** — Okta, Azure AD, Google Workspace.

    ## Rate limits
    - `/api/v1/decisions/check`: 1,000 RPS per org, burst 2,000.
    - Other endpoints: 100 RPS per principal.

    ## Fail-closed
    When the policy engine cannot reach Postgres, decisions default to `deny`
    unless the client explicitly requests degraded mode.
  contact: { name: Kynara Platform, email: platform@kynara.example.com }
  license: { name: Proprietary }
servers:
  - url: https://kynara.example.com
    description: Production
  - url: https://sandbox.kynara.example.com
    description: Sandbox
...
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /api/v1/agents:
    get:
      tags: [agents]
      summary: List agents
      responses:
        "200":
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Agent" } }
    post:
      tags: [agents]
      summary: Register a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /api/v1/agents/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [agents]
      summary: Get agent
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
    put:
      tags: [agents]
      summary: Update agent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /api/v1/agents/{id}/kill:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [agents]
      summary: Emergency kill (revokes all sessions and disables agent immediately)
      responses: { "204": { description: Killed } }
...
  /api/v1/policies:
    get:
      tags: [policies]
      summary: List policies
      responses:
        "200":
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Policy" } }
    post:
      tags: [policies]
      summary: Create policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
  /api/v1/policies/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [policies]
      summary: Update policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
    delete:
      tags: [policies]
      summary: Soft-delete policy
      responses: { "204": { description: Deleted } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Kynara-Key
...
  schemas:
...
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        slug: { type: string }
        display_name: { type: string }
        mode: { type: string, enum: [human_supervised, autonomous, read_only] }
        risk_class: { type: string, enum: [low, medium, high, critical] }
        owner_user_id: { type: string, format: uuid }
        is_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
    AgentCreate:
      type: object
      required: [slug, display_name, mode]
      properties:
        slug: { type: string }
        display_name: { type: string }
        mode: { type: string, enum: [human_supervised, autonomous, read_only] }
        risk_class: { type: string, enum: [low, medium, high, critical] }
        owner_user_id: { type: string, format: uuid }
    AgentUpdate:
      allOf:
        - $ref: "#/components/schemas/AgentCreate"
        - type: object
          properties:
            is_enabled: { type: boolean }
...
    Policy:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        effect: { type: string, enum: [allow, deny, require_approval] }
        priority: { type: integer, minimum: 0, maximum: 1000 }
        actions: { type: array, items: { type: string } }
        resource_types: { type: array, items: { type: string } }
        condition:
          description: JSON AST representing the boolean condition
          type: object
        is_enabled: { type: boolean }
    PolicyCreate:
      allOf: [{ $ref: "#/components/schemas/Policy" }]
...